Move all the file into module and start to add support for travis
