+++ .coveragerc
@@ -1,0 +1,19 @@
+# Configuration for coverage.py
+
+[run]
+branch = True
+source = unbalanced_dataset
+include = */unbalanced_dataset/*
+omit =
+    */setup.py
+
+[report]
+exclude_lines =
+    pragma: no cover
+    def __repr__
+    if self.debug:
+    if settings.DEBUG
+    raise AssertionError
+    raise NotImplementedError
+    if 0:
+    if __name__ == .__main__.:
