+++ requirements.txt
@@ -1,0 +1,3 @@
+scipy>=0.16.0
+numpy>=1.9.3
+scikit-learn>=0.17.0
