+++ setup.py
@@ -1,25 +1,128 @@
+#! /usr/bin/env python
+"""Toolbox for unbalanced dataset in machine learning."""
+
 from setuptools import setup, find_packages
+import os
+import sys
 
-install_requires = [
-    'numpy',
-    'scipy',
-    'scikit-learn',
-    ]
+import setuptools
+from distutils.command.build_py import build_py
 
-setup(name='UnbalancedDataset',
-      version='0.1',
-      description='Python module with numerous re-sampling strategies to deal '
-                  'with classification of data-sets with strong between class '
-                  'imbalance.',
-      classifiers=[
-          "Programming Language :: Python :: 2.7",
-          "Programming Language :: Python :: 3.4",
-          ],
-      author='Fernando Nogueira, Guillaume Lemaitre',
-      author_email='fmfnogueira@gmail.com, guillaume.lemaitre@udg.edu',
-      url='https://github.com/fmfn/UnbalancedDataset',
-      packages=find_packages(),
-      include_package_data=True,
-      zip_safe=False,
-      install_requires=install_requires,
-      )
+if sys.version_info[0] < 3:
+    import __builtin__ as builtins
+else:
+    import builtins
+
+descr = """Toolbox for unbalanced dataset in machine learning."""
+
+DISTNAME = 'unbalanced_dataset'
+DESCRIPTION = 'Toolbox for unbalanced dataset in machine learning.'
+LONG_DESCRIPTION = descr
+MAINTAINER = 'Fernando Nogueira, Guillaume Lemaitre'
+MAINTAINER_EMAIL = 'fmfnogueira@gmail.com, g.lemaitre58@gmail.com'
+URL = 'https://github.com/fmfn/UnbalancedDataset'
+LICENSE = 'new BSD'
+DOWNLOAD_URL = 'https://github.com/fmfn/UnbalancedDataset'
+
+# This is a bit (!) hackish: we are setting a global variable so that the main
+# skimage __init__ can detect if it is being loaded by the setup routine, to
+# avoid attempting to load components that aren't built yet:
+# the numpy distutils extensions that are used by UnbalancedDataset to 
+# recursively build the compiled extensions in sub-packages is based on
+# the Python import machinery.
+builtins.__UNBALANCED_DATASET_SETUP__ = True
+
+with open('unbalanced_dataset/__init__.py') as fid:
+    for line in fid:
+        if line.startswith('__version__'):
+            VERSION = line.strip().split()[-1][1:-1]
+            break
+
+with open('requirements.txt') as fid:
+    INSTALL_REQUIRES = [l.strip() for l in fid.readlines() if l]
+
+# requirements for those browsing PyPI
+REQUIRES = [r.replace('>=', ' (>= ') + ')' for r in INSTALL_REQUIRES]
+REQUIRES = [r.replace('==', ' (== ') for r in REQUIRES]
+REQUIRES = [r.replace('[array]', '') for r in REQUIRES]
+
+
+def configuration(parent_package='', top_path=None):
+    if os.path.exists('MANIFEST'):
+        os.remove('MANIFEST')
+
+    from numpy.distutils.misc_util import Configuration
+    config = Configuration(None, parent_package, top_path)
+
+    config.set_options(
+        ignore_setup_xxx_py=True,
+        assume_default_configuration=True,
+        delegate_options_to_subpackages=True,
+        quiet=True)
+
+    config.add_subpackage('unbalanced_dataset')
+
+    return config
+
+if __name__ == "__main__":
+    try:
+        from numpy.distutils.core import setup
+        extra = {'configuration': configuration}
+        # Do not try and upgrade larger dependencies
+        for lib in ['scipy', 'numpy', 'matplotlib']:
+            try:
+                __import__(lib)
+                INSTALL_REQUIRES = [i for i in INSTALL_REQUIRES
+                                    if lib not in i]
+            except ImportError:
+                pass
+    except ImportError:
+        if len(sys.argv) >= 2 and ('--help' in sys.argv[1:] or
+                                   sys.argv[1] in ('--help-commands',
+                                                   '--version',
+                                                   'clean')):
+            # For these actions, NumPy is not required.
+            #
+            # They are required to succeed without Numpy for example when
+            # pip is used to install UnbalancedDataset when Numpy is not yet
+            # present in the system.
+            from setuptools import setup
+            extra = {}
+        else:
+            print('To install UnbalancedDataset from source, you need numpy.' +
+                  'Install numpy with pip:\n' +
+                  'pip install numpy\n'
+                  'Or use your operating system package manager.')
+            sys.exit(1)
+
+    setup(
+        name=DISTNAME,
+        description=DESCRIPTION,
+        long_description=LONG_DESCRIPTION,
+        maintainer=MAINTAINER,
+        maintainer_email=MAINTAINER_EMAIL,
+        url=URL,
+        license=LICENSE,
+        download_url=DOWNLOAD_URL,
+        version=VERSION,
+
+        classifiers=['Intended Audience :: Science/Research',
+                     'Intended Audience :: Developers',
+                     'License :: OSI Approved',
+                     'Programming Language :: Python',
+                     'Topic :: Software Development',
+                     'Topic :: Scientific/Engineering',
+                     'Operating System :: Microsoft :: Windows',
+                     'Operating System :: POSIX',
+                     'Operating System :: Unix',
+                     'Operating System :: MacOS',
+                     'Programming Language :: Python :: 2.7',
+                     'Programming Language :: Python :: 3.4'],
+        install_requires=INSTALL_REQUIRES,
+        requires=REQUIRES,
+        packages=setuptools.find_packages(exclude=['doc']),
+        include_package_data=True,
+        zip_safe=False,  # the package can run out of an .egg file
+        cmdclass={'build_py': build_py},
+        **extra
+    )
