+++ unbalanced_dataset/combine/__init__.py
@@ -1,0 +1,9 @@
+"""The :mod:`unbalanced_dataset.combine` provides methods which combine
+over-sampling and under-sampling.
+"""
+
+from .smote_enn import SMOTEENN
+from .smote_tomek import SMOTETomek
+
+__all__ = ['SMOTEENN',
+           'SMOTETomek']
