+++ unbalanced_dataset/utils/__init__.py
@@ -1,0 +1,7 @@
+"""
+The :mod:`unbalanced_dataset.utils` provides some helper methods.
+"""
+
+from .pipeline import Pipeline
+
+__all__ = ['Pipeline']
